+++ src/main/java/com/javaweb/model/entity/person/Person.java
@@ -1,7 +1,7 @@
 package com.javaweb.model.entity.person;
 
 public class Person {
-	private long id;
+	private int id;
 	private String firstName;
 	private String secondName;
 	private Gender gender;
@@ -9,11 +9,11 @@ public class Person {
 	private String password;
 	private PersonRole role;
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 
@@ -114,7 +114,7 @@ public class Person {
 	public static class Builder {
 		private Person person = new Person();
 
-		public Builder setId(long id) {
+		public Builder setId(int id) {
 			person.setId(id);
 			return this;
 		}
